+++ packages/cli/src/services/config.ts
@@ -15,6 +15,7 @@ export enum OutputFormat {
   SARIF = 'sarif',
   CODE_CLIMATE = 'code-climate',
   GITLAB = 'gitlab',
+  MARKDOWN = 'markdown',
 }
 
 export interface ILintConfig {