+++ packages/cli/src/services/output.ts
@@ -12,6 +12,7 @@ import {
   githubActions,
   sarif,
   codeClimate,
+  markdown,
 } from '@stoplight/spectral-formatters';
 import type { Formatter, FormatterOptions } from '@stoplight/spectral-formatters';
 import type { OutputFormat } from './config';
@@ -29,6 +30,7 @@ const formatters: Record<OutputFormat, Formatter> = {
   sarif,
   'code-climate': codeClimate,
   gitlab: codeClimate,
+  markdown,
 };
 
 export function formatOutput(