+++ packages/formatters/src/index.node.ts
@@ -1,4 +1,4 @@
-export { html, json, junit, text, stylish, teamcity } from './index';
+export { html, json, junit, text, stylish, teamcity, markdown } from './index';
 export type { Formatter, FormatterOptions } from './index';
 export { pretty } from './pretty';
 export { githubActions } from './github-actions';